Updated config.fish and env.fish to include a section for login shells only, if status --is-login.
Also removed functions.fish in favor of the fish standard of a functions
directory containing <function_name>.fish for each function. I was able to
simplify the silly ssh_agent python script and fish function to just a fish
function that calls ssh-agent with the -c option, which changes bash-based
variable exports to csh-based setenv calls. fish has a defined alias for setenv
to set -gx but what is desirable is for setenv calls for ssh-agent to be set
-Ux instead so that other fish instances are able to communicate with the ssh
agent daemon.