alacritty.yml: updated shell program for mac
the env command is used instead of fish directly, so that we can specify
a desired architecture via the `arch` command (default arm64). Because
we default to arm64, we also default to the path to the fish binary
installed by the arm-based brew. Note that alacritty uses the shell
command to set the SHELL variable (as far as I can tell), so we also set
the SHELL variable in our env command